+++ app/src/main/java/com/ecarriers/drivers/models/TripLocation.java
@@ -3,8 +3,8 @@ package com.ecarriers.drivers.models;
 public class TripLocation {
 
     private long tripId;
-    private float lat;
-    private float lng;
+    private double lat;
+    private double lng;
 
     public long getTripId() {
         return tripId;
@@ -14,19 +14,19 @@ public class TripLocation {
         this.tripId = tripId;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(float lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 }
